+++ source/dynamics/orbit.hpp
@@ -81,7 +81,7 @@ class Orbit<Point<F>>
     Void insert(F t, const Point<F>& pt);
     const InterpolatedCurve& curve() const { return *this->_curve; }
   private:
-    std::shared_ptr< InterpolatedCurve > _curve;
+    std::shared_ptr<InterpolatedCurve> _curve;
 };
 
 template<class F>
@@ -92,7 +92,20 @@ public:
     Void insert(F t, const Point<Approximation<F>>& pt) { this->_curve->insert(t,pt); }
     const InterpolatedCurve& curve() const { return *this->_curve; }
 private:
-    std::shared_ptr< InterpolatedCurve > _curve;
+    std::shared_ptr<InterpolatedCurve> _curve;
+};
+
+
+template<class F>
+class Orbit<Vector<Point<Approximation<F>>>>
+{
+  public:
+    Orbit(const Vector<Point<Approximation<F>>>& ptLst);
+    Void insert(F t, const Point<Approximation<F>>& pt, SizeType curveNumber);
+    const Vector<InterpolatedCurve>& curves() { return *_curves; }
+    const InterpolatedCurve& curve(SizeType i) { return _curves->at(i); }
+  private:
+    std::shared_ptr<Vector<InterpolatedCurve>> _curves;
 };
 
 template<class F>
@@ -102,11 +115,34 @@ class Orbit<LabelledPoint<Approximation<F>>>
   public:
     Orbit(const LabelledPoint<Approximation<F>>& pt) : _curve(new LabelledInterpolatedCurve(pt)) { }
     Void insert(F t, const LabelledPoint<Approximation<F>>& pt) { this->_curve->insert(t,pt); }
-    const LabelledInterpolatedCurve& curve() const { return *this->_curve; }
+    const LabelledInterpolatedCurve& curves() const { return *this->_curve; }
     virtual Void draw(CanvasInterface& canvas, const Variables2d& axes) const override { _curve->draw(canvas,axes); }
     virtual LabelledDrawable2dInterface* clone() const override { return new Orbit<LabelledPoint<Approximation<F>>>(*this); }
   private:
-    std::shared_ptr< LabelledInterpolatedCurve > _curve;
+    std::shared_ptr<LabelledInterpolatedCurve> _curve;
+};
+
+template<class F>
+class Orbit<Vector<LabelledPoint<Approximation<F>>>>
+    : public LabelledDrawable2dInterface
+{
+  public:
+    Orbit(const Vector<LabelledPoint<Approximation<F>>>& ptLst)
+        : _curves(new Vector(ptLst.size(), LabelledInterpolatedCurve(ptLst.at(0)))) {
+        for (SizeType i=1; i<ptLst.size(); i++)
+            _curves->at(i) = LabelledInterpolatedCurve(ptLst.at(i));
+    }
+    Void insert(F t, const LabelledPoint<Approximation<F>>& pt, SizeType curveNumber){
+        _curves->at(curveNumber).insert(static_cast<const RawFloatDP>(t), pt);
+    }
+    const Vector<LabelledInterpolatedCurve>& curves() {
+      return *_curves;
+    }
+    const LabelledInterpolatedCurve& curve(SizeType i) { return _curves->at(i); }
+    virtual Void draw(CanvasInterface& canvas, const Variables2d& axes) const override { for (SizeType i=0; i<_curves->size(); i++){ _curves->at(i).draw(canvas, axes); } }
+    virtual LabelledDrawable2dInterface* clone() const override { return new Orbit<Vector<LabelledPoint<Approximation<F>>>>(*this); }
+  private:
+    std::shared_ptr<Vector<LabelledInterpolatedCurve>> _curves;
 };
 
 template<>
@@ -128,6 +164,14 @@ class Orbit<Storage>
     std::shared_ptr<Data> _data;
 };
 
+struct Orbit<Storage>::Data {
+    Data(const Grid& grid, const EffectiveVectorMultivariateFunction& auxiliary_mapping)
+            : initial(grid,auxiliary_mapping), reach(initial), intermediate(initial), final(initial) { }
+    Storage initial;
+    Storage reach;
+    Storage intermediate;
+    Storage final;
+};
 
 template<>
 class Orbit<Enclosure>