+++ source/dynamics/vector_field_simulator.hpp
@@ -1,7 +1,7 @@
 /***************************************************************************
  *            dynamics/vector_field_simulator.hpp
  *
- *  Copyright  2009-21  Luca Geretti
+ *  Copyright  2009-21  Luca Geretti, Mirko Albanese
  *
  ****************************************************************************/
 
@@ -35,10 +35,27 @@
 #include "geometry/point.hpp"
 #include "solvers/configuration_interface.hpp"
 #include "dynamics/vector_field.hpp"
+#include "dynamics/orbit.hpp"
+
+#include "betterthreads/workload.hpp"
 
 using namespace ConcLog;
 
 namespace Ariadne {
+
+enum class DiscretisationType
+{
+  Mince,
+  Recombine
+};
+
+OutputStream& operator<<(OutputStream& os, const DiscretisationType& dtype) {
+
+    if(dtype == DiscretisationType::Mince){ os << "Mince"; }
+    else{ os << "Recombine"; }
+    
+    return os;
+}
 
 class VectorField;
 class VectorFieldSimulatorConfiguration;
@@ -53,13 +70,24 @@ class VectorFieldSimulator
 {
   public:
     typedef LabelledPoint<FloatDPApproximation> ApproximatePointType;
+    typedef Vector<ApproximatePointType> ApproximateListPointType;
     typedef LabelledPoint<Real> RealPointType;
+    typedef Vector<RealPointType> RealListPointType;
     typedef RealVariablesBox RealBoxType;
     typedef Real TerminationType;
     typedef VectorField SystemType;
     typedef VectorFieldSimulatorConfiguration ConfigurationType;
-    typedef Orbit<ApproximatePointType> OrbitType;
+    typedef Orbit<ApproximateListPointType> OrbitType;
 
+    //! \brief Synchronised wrapping of orbit to allow concurrent adjoining
+    struct SynchronisedOrbit : public OrbitType {
+        SynchronisedOrbit(ApproximateListPointType const& initial_points) : OrbitType(initial_points) { }
+        void insert(FloatDP const& t, ApproximatePointType const& pt, SizeType const& curve_number) {
+            BetterThreads::LockGuard<BetterThreads::Mutex> lock(_mux); OrbitType::insert(t,pt,curve_number); }
+      private:
+        BetterThreads::Mutex _mux;
+    };
+    typedef BetterThreads::StaticWorkload<Pair<SizeType,ApproximatePointType>, TerminationType const&, SharedPointer<SynchronisedOrbit>> WorkloadType;
   public:
 
     //! \brief Default constructor.
@@ -74,13 +102,15 @@ class VectorFieldSimulator
     //!@{
     //! \name Simulation using points.
     //! \brief Compute an approximation to the orbit set.
-    OrbitType orbit(ApproximatePointType const& initial_point, TerminationType const& termination) const;
-    OrbitType orbit(RealPointType const& initial_point, TerminationType const& termination) const;
     OrbitType orbit(RealBoxType const& initial_box, TerminationType const& termination) const;
     OrbitType orbit(RealExpressionBoundedConstraintSet const& initial_set, TerminationType const& termination) const;
+    OrbitType orbit(ApproximateListPointType const& initial_list, TerminationType const& termination) const;
+    OrbitType orbit(UpperBoxType& initial_box, TerminationType const& termination) const;
 
   private:
 
+    void _simulate_from_point(Pair<SizeType,ApproximatePointType> indexed_initial, TerminationType const& termination, SharedPointer<SynchronisedOrbit> orbit) const;
+  private:
     SharedPointer<SystemType> _system;
     SharedPointer<ConfigurationType> _configuration;
 };
@@ -99,11 +129,19 @@ class VectorFieldSimulatorConfiguration : public ConfigurationInterface
 
     //! \brief The fixed integration step size to use.
     FloatDPApproximation _step_size;
+    Nat _num_subdivisions;
+    DiscretisationType _discretisation_type;
 
   public:
 
     Void set_step_size(double h) { _step_size=h; }
     FloatDPApproximation const& step_size() const { return _step_size; }
+
+    Void set_discretisation_type(DiscretisationType discretisation_type) { _discretisation_type = discretisation_type; }
+    DiscretisationType const& discretisation_type() const { return _discretisation_type; }
+
+    Void set_num_subdivisions(Nat num_subdivisions) { _num_subdivisions = num_subdivisions; }
+    Nat const& num_subdivisions() const { return _num_subdivisions; }
 
     virtual OutputStream& _write(OutputStream& os) const;